Refactor Code Favourite and  getcategory and getsubcategory
