+++ myapp/app.js
@@ -1,29 +1,22 @@
 const mysql = require("mysql");
 const express = require("express");
-const session = require("express-session");
+const sessions = require("express-session");
 const path = require("path");
 var connection = require("./database/config.js");
-
-var registrationRouter = require("./routes/registration-route.js");
-var loginRouter = require("./routes/login-route.js");
-var dashboardRouter = require("./routes/dashboard-route.js");
-var admindashboardRouter = require("./routes/admindashboard-route.js");
-var logoutRouter = require("./routes/logout-route.js");
-var productRouter = require("./routes/product-route.js");
-var categoryRouter = require("./routes/category-route.js");
-var subcategoryRouter = require("./routes/subcategory-route.js");
-var searchRouter = require("./routes/search-route.js");
-var favouriteRouter = require("./routes/favourite-route.js");
-var mailRouter = require("./routes/mail-route.js");
+var appRouter = require("./routes/route.js");
 
 const app = express();
-app.use(
-  session({
-    secret: "secret",
-    resave: true,
-    saveUninitialized: true,
-  })
-);
+//Session Middleware
+
+const oneDay = 1000 * 60 * 60 * 24;
+
+//session middleware
+app.use(sessions({
+    secret: "thisismysecrctekeyfhrgfgrfrty84fwir767",
+    saveUninitialized:true,
+    cookie: { maxAge: oneDay },
+    resave: false
+}));
 
 app.use(express.json());
 app.use(express.urlencoded({ extended: true }));
@@ -32,18 +25,8 @@ app.use(express.static(path.join(__dirname, "public")));
 // view engine setup
 app.set("views", path.join(__dirname, "views"));
 app.set("view engine", "ejs");
+app.use("/", appRouter);
 
-app.use("/", registrationRouter);
-app.use("/", loginRouter);
-app.use("/", dashboardRouter);
-app.use("/", admindashboardRouter);
-app.use("/", logoutRouter);
-app.use("/", productRouter);
-app.use("/", categoryRouter);
-app.use("/", subcategoryRouter);
-app.use("/", searchRouter);
-app.use("/", favouriteRouter);
-app.use("/", mailRouter);
 
 
 app.listen(3000, () => {
