+++ myapp/controller/all_product.js
@@ -21,7 +21,7 @@ function Productshow(req,res){
 function Create(req, res, next) {
     var product_name= req.body.pname;
    
-     sql = `INSERT INTO Product (product_name) VALUES ("${product_name}")`;
+    sql = `INSERT INTO Product (product_name) VALUES ("${product_name}")`;
     db.query(sql, function(err, result) {
       if (err) throw err;
       console.log('record inserted');  
@@ -47,7 +47,6 @@ function Create(req, res, next) {
     var sql=`SELECT * FROM Product WHERE pid=${productId}`;
     db.query(sql, function (err, data) {
       if (err) throw err;
-     
       res.render('product/edit', {editData: data[0]});
     });
   }
