--- myapp/controller/mail.js
@@ -1,31 +1,0 @@
-const nodemailer = require('nodemailer');
-const mailGun = require('nodemailer-mailgun-transport');
-
-const auth = {
-    auth: {
-        api_key: 'put-your-API-key',
-        domain: 'put-your-sand-box-domain-from-mailgun'
-    }
-};
-
-const transporter = nodemailer.createTransport(mailGun(auth));
-
-const sendMail = (name, email, subject, text, cb) => {
-    const mailOptions = {
-        sender: name,
-        from: email,
-        to: 'gourpriya20@gmail.com',
-        subject: subject,
-        text: text
-    };
-    transporter.sendMail(mailOptions, function(err, data) {
-        if (err) {
-            cb(err, null);
-        } else {
-            cb(null, data);
-        }
-    });
-}
-
-// Exporting the sendmail
-module.exports = sendMail;
