--- myapp/routes/admindashboard-route.js
@@ -1,12 +1,0 @@
-var express = require("express");
-var router = express.Router();
-/* GET users listing. */
-
-router.get("/admindashboard", function (req, res, next) {
-  if (req.session.loggedinUser) {
-    res.render("admindashboard", { email: req.session.emailAddress });
-  } else {
-    res.redirect("/login");
-  }
-});
-module.exports = router;
