--- myapp/routes/dashboard-route.js
@@ -1,11 +1,0 @@
-var express = require("express");
-var router = express.Router();
-/* GET users listing. */
-router.get("/dashboard", function (req, res, next) {
-  if (req.session.loggedinUser) {
-    res.render("dashboard", { email: req.session.emailAddress });
-  } else {
-    res.redirect("/login");
-  }
-});
-module.exports = router;
