--- myapp/routes/logout-route.js
@@ -1,8 +1,0 @@
-var express = require("express");
-var router = express.Router();
-/* GET users listing. */
-router.get("/logout", function (req, res) {
-  req.session.destroy();
-  res.redirect("/login");
-});
-module.exports = router;
