+++ myapp/views/favourite/create.ejs
@@ -1,0 +1,21 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/css/bootstrap.min.css">
+    <title>Product</title>
+</head>
+<body>
+    <%- include('../header.ejs') %>
+    <div>
+      <form action="<%=(typeof editData!='undefined')?'/favourite-add/'+editData.sid:'/favourite/wishlist'%>" method="POST">
+        <label>SubCategory ID</label>
+        <input type="text" placeholder="SubCategory Id" name="sid" required value="<%=(typeof editData!='undefined')? editData.sid:''%>">
+        <button type="submit">Submit</button>
+      </form>
+  </div>
+
+</body>
+</html>
