+++ myapp/views/favourite/wishlist.ejs
@@ -17,8 +17,6 @@
             <th>S.N</th>
             <th>SubCategory ID</th>
             <th>DELETE</th>
-         
-           
         </tr>
         
         <%
