+++ myapp/views/getSubcategory.ejs
@@ -31,8 +31,8 @@
         <tr>
             <td><%=i; %></td>
             <td><%=data.Description %></td>
-            <td><%=data.price%></td>
-            <td><%=data.cid%></td>
+            <td><%=data.price %></td>
+            <td><%=data.cid %></td>
             <td><%=data.product_name %></td>
         
         </tr>
