--- myapp/views/mail.ejs
@@ -1,81 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Mailgun App</title>
-    <link href="https://stackpath.bootstrapcdn.com/bootstrap/4.5.0/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-9aIt2nRpC12Uk9gS9baDl411NQApFmC26EwAOH8WgZl5MYYxFfc+NcPb1dKGj7Sk" crossorigin="anonymous">
-    <style>
-        body {
-            box-sizing: border-box;
-        }
-
-        form {
-            /* margin: 10%; */
-            width: 400px;
-            /* margin: 10% 40%; */
-        }
-
-        .container {
-            margin: 10% 30%;
-        }
-
-        textarea {
-            height: 300px;
-        }
-    </style>
-</head>
-
-<body>
-    <div class="container">
-        <form>
-            <div class="form-group">
-                <label for="exampleInputEmail1">Email address</label>
-                <input type="email" class="form-control" id="email" aria-describedby="emailHelp" placeholder="Enter email" required>
-                <small id="emailHelp" class="form-text text-muted">We'll never share your email with anyone else.</small>
-            </div>
-            <div class="form-group">
-                <label for="exampleInputEmail1">Subject</label>
-                <input type="text" class="form-control" id="subject" aria-describedby="emailHelp" placeholder="Subject" required>
-            </div>
-            <div class="form-group">
-                <label for="exampleInputPassword1">Message</label>
-                <textarea class="form-control" id="text" placeholder="Message" required></textarea>
-            </div>
-            <button type="submit" class="btn btn-primary">Submit</button>
-        </form>
-    </div>
-
-    <!--  -->
-    <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.5.0/js/bootstrap.bundle.min.js"></script>
-
-    <!-- Here is a script where we will be collecting data from the form  -->
-    <script>
-        $('form').on('submit', (e) => {
-            // e.preventDefault();
-
-            const name = $('#name').val().trim();
-            const email = $('#from-email').val().trim();
-            const subject = $('#subject').val().trim();
-            const text = $('#text').val().trim();
-
-            const data = {
-                name,
-                email,
-                subject,
-                text
-            };
-
-            $.post('/email', data, function() {
-                console.log('Server recieved our data');
-
-
-            });
-        });
-        // document.getElementById('#form-data').reset();
-    </script>
-
-</body>
-
-</html>
