+++ myapp/views/subcategory/edit.ejs
@@ -8,7 +8,7 @@
     <title>Update Category</title>
 </head>
 <body>
-     <!-- <%- include('../adminheader.ejs') %>  -->
+     <%- include('../adminheader.ejs') %> 
      <div>
